pyproject.toml
[project]
name = "gigwise-analytics"
version = "0.1.0"
description = "venv for DE Zoomcamp 2026 project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "dbt-bigquery>=1.11.1",
    "dbt-core>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "dlt[duckdb,workspace]>=1.22.1",
    "duckdb>=1.4.4",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.33.1",
    "google-cloud-storage>=3.1.1",
    "ibis-framework[duckdb]>=12.0.0",
    "ipykernel>=7.2.0",
    "kafka-python>=2.3.0",
    "marimo>=0.20.2",
    "numpy>=2.4.2",
    "pandas>=2.2,<3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pyspark>=3.5,<4",
    "streamlit>=1.55.0",
]
